rename viewModel to Provider
