+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:mboacare/colors.dart';
-import 'package:mboacare/sign_up/view_model/signup_view_model.dart';
+import 'package:mboacare/sign_up/view_model/signup_provider.dart';
 import 'package:mboacare/splash.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:mboacare/utils/constants.dart';
@@ -11,7 +11,7 @@ import 'hospital_provider.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'locale_provider.dart';
 import 'l10n/app_localizations.dart';
-import 'login/signin_view_model.dart';
+import 'login/signin_provider.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
